requirements.txt
streamlit==1.35.0
pdfplumber==0.10.4
python-docx==1.1.2
pandas==2.2.1
